+++ scripts/testCodeGenWithFunctional.ts
@@ -71,6 +71,8 @@ const main = () => {
   Writer.generateParameter("test/infer.domain/index.yml", "test/code/functional/parameter/infer.domain.json");
 
   Writer.generateFormatTypeCode("test/format.domain/index.yml", "test/code/functional/format.domain/code.ts");
+
+  Writer.generateFormatTypeCode("test/cloudflare/openapi.yaml", "test/code/functional/cloudflare/client.ts");
 };
 
 main();
